+++ web-client/src/tests/journey/petitionsClerkLogIn.js
@@ -1,0 +1,10 @@
+export default test => {
+  return it('Petitions clerk logs in', async () => {
+    test.setState('user', {
+      name: 'Petitions Clerk',
+      role: 'petitionsclerk',
+      token: 'petitionsclerk',
+      userId: 'petitionsclerk',
+    });
+  });
+};
