+++ web-client/src/tests/journey/petitionsClerkViewsDashboard.js
@@ -1,0 +1,7 @@
+export default test => {
+  return it('Petitions clerk views dashboard', async () => {
+    await test.runSequence('gotoDashboardSequence');
+    expect(test.getState('currentPage')).toEqual('DashboardPetitionsClerk');
+    expect(test.getState('cases').length).toBeGreaterThan(0);
+  });
+};
